Вынес функции сортировки и константы в отдельный модуль
